edited schema script as amount type was incorrect (I think) and then added sql statements to repositories ... still need to work out how I create an account object

+++ src/main/java/budgetApp/classes/Income.java
@@ -9,7 +9,7 @@ public class Income implements Transaction {
     private int accountId;
 
 
-    public Income(int id, String type, String description, int amount, int accountId) {
+    public Income(int id, String type, String description, double amount, int accountId) {
         this.id = id;
         this.type = type;
         this.description = description;
